backend/scripts/validate_apis.py
#!/usr/bin/env python3
"""
Simple API Validation Script
Tests all configured APIs and displays results
"""

import os
import sys
import json
import requests
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Color codes for terminal output
GREEN = '\033[92m'
YELLOW = '\033[93m'
RED = '\033[91m'
BLUE = '\033[94m'
RESET = '\033[0m'

def print_header(text):
    print(f"\n{BLUE}{'='*70}{RESET}")
    print(f"{BLUE}{text:^70}{RESET}")
    print(f"{BLUE}{'='*70}{RESET}\n")

def print_success(text):
    print(f"{GREEN}✅ {text}{RESET}")

def print_warning(text):
    print(f"{YELLOW}⚠️  {text}{RESET}")

def print_error(text):
    print(f"{RED}❌ {text}{RESET}")

def test_weather_gov():
    """Test Weather.gov API"""
    try:
        user_agent = os.getenv('WEATHER_GOV_USER_AGENT', '')
        if not user_agent:
            print_warning("Weather.gov: No user agent configured")
            return False

        headers = {'User-Agent': user_agent}
        response = requests.get(
            'https://api.weather.gov/alerts/active?area=MI',
            headers=headers,
            timeout=10,
            verify=False
        )

        if response.status_code == 200:
            data = response.json()
            alerts = data.get('features', [])
            print_success(f"Weather.gov: {len(alerts)} active alerts in Michigan")
            if alerts:
                for alert in alerts[:2]:
                    event = alert['properties']['event']
                    area = alert['properties']['areaDesc']
                    print(f"    - {event} in {area}")
            return True
        else:
            print_error(f"Weather.gov: HTTP {response.status_code}")
            return False

    except Exception as e:
        print_error(f"Weather.gov: {str(e)}")
        return False

def test_noaa_api():
    """Test NOAA Storm Events API"""
    try:
        token = os.getenv('NOAA_API_TOKEN', '')
        if not token or token == 'your-noaa-token-here':
            print_warning("NOAA: No API token configured")
            return False

        headers = {'token': token}
        response = requests.get(
            'https://www.ncei.noaa.gov/cdo-web/api/v2/datasets?limit=3',
            headers=headers,
            timeout=10,
            verify=False
        )

        if response.status_code == 200:
            data = response.json()
            datasets = data.get('results', [])
            print_success(f"NOAA API: Token valid, {len(datasets)} datasets available")

            # Test storm data for Oakland County, MI
            response = requests.get(
                'https://www.ncei.noaa.gov/cdo-web/api/v2/data',
                params={
                    'datasetid': 'GHCND',
                    'locationid': 'FIPS:26125',  # Oakland County
                    'startdate': '2024-01-01',
                    'enddate': '2024-12-31',
                    'limit': 5,
                    'datatypeid': 'PRCP'
                },
                headers=headers,
                timeout=10,
                verify=False
            )

            if response.status_code == 200:
                storm_data = response.json()
                records = storm_data.get('results', [])
                print(f"    - Oakland County: {len(records)} weather records found")

            return True
        else:
            print_error(f"NOAA API: HTTP {response.status_code}")
            return False

    except Exception as e:
        print_error(f"NOAA API: {str(e)}")
        return False

def test_google_maps():
    """Test Google Maps Geocoding API"""
    try:
        api_key = os.getenv('GOOGLE_MAPS_API_KEY', '')
        if not api_key or api_key == 'your-google-api-key-here':
            print_warning("Google Maps: No API key configured")
            return False

        response = requests.get(
            'https://maps.googleapis.com/maps/api/geocode/json',
            params={
                'address': '1234 Main St, Bloomfield Hills, MI',
                'key': api_key
            },
            timeout=10,
            verify=False
        )

        if response.status_code == 200:
            data = response.json()
            status = data.get('status')

            if status == 'OK':
                results = data.get('results', [])
                print_success(f"Google Maps: Geocoding working, {len(results)} results")
                if results:
                    location = results[0]['geometry']['location']
                    address = results[0]['formatted_address']
                    print(f"    - Test address: {address}")
                    print(f"    - Coordinates: {location['lat']:.6f}, {location['lng']:.6f}")
                return True
            elif status == 'REQUEST_DENIED':
                error_msg = data.get('error_message', 'Unknown error')
                print_error(f"Google Maps: {error_msg}")
                print(f"    - Make sure Geocoding API is enabled in Google Cloud Console")
                return False
            else:
                print_error(f"Google Maps: Status={status}")
                return False
        else:
            print_error(f"Google Maps: HTTP {response.status_code}")
            return False

    except Exception as e:
        print_error(f"Google Maps: {str(e)}")
        return False

def test_zillow():
    """Test Zillow API"""
    api_key = os.getenv('ZILLOW_API_KEY', '')
    if not api_key or api_key == 'your-zillow-key-here':
        print_warning("Zillow: No API key configured (registration pending)")
        print("    - Register at: https://www.zillowgroup.com/developers/")
        return False

    print_warning("Zillow: API key configured but not tested (requires approval)")
    return False

def test_twitter():
    """Test Twitter API"""
    bearer_token = os.getenv('TWITTER_BEARER_TOKEN', '')
    if not bearer_token or bearer_token == 'your-twitter-token-here':
        print_warning("Twitter: No bearer token configured")
        print("    - Register at: https://developer.twitter.com/")
        return False

    print_warning("Twitter: API configured but not tested")
    return False

def test_facebook():
    """Test Facebook API"""
    access_token = os.getenv('FACEBOOK_ACCESS_TOKEN', '')
    if not access_token or access_token == 'your-facebook-token-here':
        print_warning("Facebook: No access token configured")
        print("    - Register at: https://developers.facebook.com/")
        return False

    print_warning("Facebook: API configured but not tested")
    return False

def main():
    """Run all API tests"""
    print_header("iSwitch Roofs CRM - API Validation Report")
    print(f"Timestamp: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n")

    results = {
        'timestamp': datetime.now().isoformat(),
        'tests': {}
    }

    # Test each API
    print_header("Priority 1: FREE APIs (Critical for Lead Generation)")

    results['tests']['weather_gov'] = test_weather_gov()
    results['tests']['noaa'] = test_noaa_api()
    results['tests']['google_maps'] = test_google_maps()

    print_header("Priority 2: Property & Social APIs")

    results['tests']['zillow'] = test_zillow()
    results['tests']['twitter'] = test_twitter()
    results['tests']['facebook'] = test_facebook()

    # Summary
    print_header("Summary")

    working = sum(1 for v in results['tests'].values() if v is True)
    pending = sum(1 for v in results['tests'].values() if v is False)
    total = len(results['tests'])

    print(f"Working APIs: {GREEN}{working}/{total}{RESET}")
    print(f"Pending Setup: {YELLOW}{pending}/{total}{RESET}\n")

    if working >= 3:
        print_success("✅ READY FOR LEAD GENERATION")
        print(f"{GREEN}With {working} working APIs, you can generate real leads now!{RESET}\n")
        print("Next steps:")
        print("  1. Start backend: cd backend && python3 main.py")
        print("  2. Start frontend: cd frontend-streamlit && streamlit run Home.py")
        print("  3. Navigate to: Live Data Generator")
        print("  4. Click: 'Generate 100 Real Leads'\n")

        print("Expected lead sources:")
        if results['tests'].get('weather_gov'):
            print("  - Weather.gov: 50-100 alert-based leads/month")
        if results['tests'].get('noaa'):
            print("  - NOAA Storms: 100-200 storm-damaged leads/month")
        if results['tests'].get('google_maps'):
            print("  - Google Maps: Address validation for all leads")

        print(f"\n{GREEN}Estimated revenue potential: $1.8M-$3.0M/month{RESET}")
    else:
        print_warning("⚠️  PARTIAL CONFIGURATION")
        print("Configure more APIs to increase lead volume\n")

    # Save results
    report_file = f"reports/api_validation_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json"
    os.makedirs('reports', exist_ok=True)

    with open(report_file, 'w') as f:
        json.dump(results, f, indent=2)

    print(f"Report saved: {report_file}\n")

    return 0 if working >= 3 else 1

if __name__ == '__main__':
    sys.exit(main())
