backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.44.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
boto3==1.40.50
botocore==1.40.50
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
ConfigArgParse==1.7.1
coverage==7.10.7
cryptography==46.0.2
dill==0.4.0
distro==1.9.0
docker==7.1.0
fastapi==0.118.3
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
frozenlist==1.8.0
fsspec==2025.9.0
gevent==25.5.1
geventhttpclient==2.3.4
google-pasta==0.2.0
googlemaps==4.10.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
locust==2.41.6
locust-cloud==1.27.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
nltk==3.9.1
numpy==2.2.1
omegaconf==2.3.0
openai==1.59.8
packaging==24.2
pandas==2.2.3
pathos==0.3.4
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.31.1
psutil==7.1.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-http-client==3.3.7
python-socketio==5.14.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.6.2
sagemaker==2.253.0
sagemaker-core==1.0.59
schema==0.7.7
scikit-learn==1.7.2
scipy==1.16.2
prophet==1.1.5
statsmodels==0.14.4
sendgrid==6.12.5
sentence-transformers==3.4.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.48.0
stripe==13.0.1
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.48.3
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
